auth_service/requirements.txt
Flask==3.1.1
Flask-SQLAlchemy==3.1.1
Flask-CORS==6.0.0
PyJWT==2.10.1
bcrypt==4.3.0
requests==2.32.3 